Added aggregate types IGS and INC; Moved enums to different project/namespace.
