+++ BlueTracker.SDK.Performance/Enums/AggregateOptions.cs
@@ -1,6 +1,7 @@
 using System.Runtime.Serialization;
+using Newtonsoft.Json;
 
-namespace BlueTracker.SDK.Performance.Report
+namespace BlueTracker.SDK.Performance.Enums
 {
     /// <summary>
     /// Types of aggregates.
@@ -11,18 +12,35 @@ namespace BlueTracker.SDK.Performance.Report
         /// Main engine of a vessel (used for propulsion), which may be fitted with a shaft generator.
         /// </summary>
         [EnumMember(Value = "mainEngine")]
+        [JsonProperty(PropertyName = "mainEngine")]
         MainEngine,
 
         /// <summary>
         /// Auxilliary engine of a vessel (used for electrical power generation).
         /// </summary>
         [EnumMember(Value = "auxEngine")]
+        [JsonProperty(PropertyName = "auxEngine")]
         AuxEngine,
 
         /// <summary>
         /// Auxilliary boiler of a vessel (with own fuel burner, no exhaust gas boiler).
         /// </summary>
         [EnumMember(Value = "boiler")]
-        Boiler
+        [JsonProperty(PropertyName = "boiler")]
+        Boiler,
+
+        /// <summary>
+        /// Inert Gas Generator
+        /// </summary>
+        [EnumMember(Value = "inertGasGenerator")]
+        [JsonProperty(PropertyName = "inertGasGenerator")]
+        InertGasGenerator,
+
+        /// <summary>
+        /// Incinerator
+        /// </summary>
+        [EnumMember(Value = "incinerator")]
+        [JsonProperty(PropertyName = "incinerator")]
+        Incinerator
     }
 }
