+++ BlueTracker.SDK.Performance/Enums/OnboardSampleSourceType.cs
@@ -1,9 +1,18 @@
 ﻿namespace BlueTracker.SDK.Performance.Enums
 {
+    /// <summary>
+    /// Types of onboard sample sources.
+    /// </summary>
     public enum OnboardSampleSourceType
     {
+        /// <summary>
+        /// A sample source of unspecific type.
+        /// </summary>
         Unspecified = 0,
 
+        /// <summary>
+        /// A sample source for measured data.
+        /// </summary>
         Measured = 1
     }
 }
