+++ BlueTracker.SDK.Performance/Post/EventData.cs
@@ -1,4 +1,5 @@
 using System;
+using BlueTracker.SDK.Performance.Enums;
 using BlueTracker.SDK.Performance.Report;
 using Newtonsoft.Json;
