+++ BlueTracker.SDK.Performance/Query/Event.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using BlueTracker.SDK.Performance.Report;
+using BlueTracker.SDK.Performance.Enums;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
