+++ BlueTracker.SDK.Performance/Query/Ship.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using BlueTracker.SDK.Performance.Enums;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
