+++ BlueTracker.SDK.Performance/Report/AggregateConsumption.cs
@@ -1,4 +1,5 @@
-﻿using Newtonsoft.Json;
+﻿using BlueTracker.SDK.Performance.Enums;
+using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 
 namespace BlueTracker.SDK.Performance.Report
