+++ BlueTracker.SDK.Performance/Report/PerformanceReport.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using BlueTracker.SDK.Performance.Enums;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
